requirements.txt
yfinance
streamlit
google-genai
google.generativeai
dotenv